pyproject.toml
[project]
name = "ml2025f-project-2"
version = "0.1.0"
description = "AI CUP 2025秋季賽 - 電腦斷層心臟肌肉影像分割競賽 II - 主動脈瓣物件偵測"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=1.1.4",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.3.226",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
